Trait rusoto_alexaforbusiness::AlexaForBusiness

pub trait AlexaForBusiness {
    fn approve_skill(
        &self, 
        input: ApproveSkillRequest
    ) -> RusotoFuture<ApproveSkillResponse, ApproveSkillError>;

    fn associate_contact_with_address_book(
        &self, 
        input: AssociateContactWithAddressBookRequest
    ) -> RusotoFuture<AssociateContactWithAddressBookResponse, AssociateContactWithAddressBookError>;

    fn associate_device_with_room(
        &self, 
        input: AssociateDeviceWithRoomRequest
    ) -> RusotoFuture<AssociateDeviceWithRoomResponse, AssociateDeviceWithRoomError>;

    fn associate_skill_group_with_room(
        &self, 
        input: AssociateSkillGroupWithRoomRequest
    ) -> RusotoFuture<AssociateSkillGroupWithRoomResponse, AssociateSkillGroupWithRoomError>;

    fn associate_skill_with_skill_group(
        &self, 
        input: AssociateSkillWithSkillGroupRequest
    ) -> RusotoFuture<AssociateSkillWithSkillGroupResponse, AssociateSkillWithSkillGroupError>;

    fn associate_skill_with_users(
        &self, 
        input: AssociateSkillWithUsersRequest
    ) -> RusotoFuture<AssociateSkillWithUsersResponse, AssociateSkillWithUsersError>;

    fn create_address_book(
        &self, 
        input: CreateAddressBookRequest
    ) -> RusotoFuture<CreateAddressBookResponse, CreateAddressBookError>;

    fn create_business_report_schedule(
        &self, 
        input: CreateBusinessReportScheduleRequest
    ) -> RusotoFuture<CreateBusinessReportScheduleResponse, CreateBusinessReportScheduleError>;

    fn create_conference_provider(
        &self, 
        input: CreateConferenceProviderRequest
    ) -> RusotoFuture<CreateConferenceProviderResponse, CreateConferenceProviderError>;

    fn create_contact(
        &self, 
        input: CreateContactRequest
    ) -> RusotoFuture<CreateContactResponse, CreateContactError>;

    fn create_profile(
        &self, 
        input: CreateProfileRequest
    ) -> RusotoFuture<CreateProfileResponse, CreateProfileError>;

    fn create_room(
        &self, 
        input: CreateRoomRequest
    ) -> RusotoFuture<CreateRoomResponse, CreateRoomError>;

    fn create_skill_group(
        &self, 
        input: CreateSkillGroupRequest
    ) -> RusotoFuture<CreateSkillGroupResponse, CreateSkillGroupError>;

    fn create_user(
        &self, 
        input: CreateUserRequest
    ) -> RusotoFuture<CreateUserResponse, CreateUserError>;

    fn delete_address_book(
        &self, 
        input: DeleteAddressBookRequest
    ) -> RusotoFuture<DeleteAddressBookResponse, DeleteAddressBookError>;

    fn delete_business_report_schedule(
        &self, 
        input: DeleteBusinessReportScheduleRequest
    ) -> RusotoFuture<DeleteBusinessReportScheduleResponse, DeleteBusinessReportScheduleError>;

    fn delete_conference_provider(
        &self, 
        input: DeleteConferenceProviderRequest
    ) -> RusotoFuture<DeleteConferenceProviderResponse, DeleteConferenceProviderError>;

    fn delete_contact(
        &self, 
        input: DeleteContactRequest
    ) -> RusotoFuture<DeleteContactResponse, DeleteContactError>;

    fn delete_device(
        &self, 
        input: DeleteDeviceRequest
    ) -> RusotoFuture<DeleteDeviceResponse, DeleteDeviceError>;

    fn delete_profile(
        &self, 
        input: DeleteProfileRequest
    ) -> RusotoFuture<DeleteProfileResponse, DeleteProfileError>;

    fn delete_room(
        &self, 
        input: DeleteRoomRequest
    ) -> RusotoFuture<DeleteRoomResponse, DeleteRoomError>;

    fn delete_room_skill_parameter(
        &self, 
        input: DeleteRoomSkillParameterRequest
    ) -> RusotoFuture<DeleteRoomSkillParameterResponse, DeleteRoomSkillParameterError>;

    fn delete_skill_authorization(
        &self, 
        input: DeleteSkillAuthorizationRequest
    ) -> RusotoFuture<DeleteSkillAuthorizationResponse, DeleteSkillAuthorizationError>;

    fn delete_skill_group(
        &self, 
        input: DeleteSkillGroupRequest
    ) -> RusotoFuture<DeleteSkillGroupResponse, DeleteSkillGroupError>;

    fn delete_user(
        &self, 
        input: DeleteUserRequest
    ) -> RusotoFuture<DeleteUserResponse, DeleteUserError>;

    fn disassociate_contact_from_address_book(
        &self, 
        input: DisassociateContactFromAddressBookRequest
    ) -> RusotoFuture<DisassociateContactFromAddressBookResponse, DisassociateContactFromAddressBookError>;

    fn disassociate_device_from_room(
        &self, 
        input: DisassociateDeviceFromRoomRequest
    ) -> RusotoFuture<DisassociateDeviceFromRoomResponse, DisassociateDeviceFromRoomError>;

    fn disassociate_skill_from_skill_group(
        &self, 
        input: DisassociateSkillFromSkillGroupRequest
    ) -> RusotoFuture<DisassociateSkillFromSkillGroupResponse, DisassociateSkillFromSkillGroupError>;

    fn disassociate_skill_from_users(
        &self, 
        input: DisassociateSkillFromUsersRequest
    ) -> RusotoFuture<DisassociateSkillFromUsersResponse, DisassociateSkillFromUsersError>;

    fn disassociate_skill_group_from_room(
        &self, 
        input: DisassociateSkillGroupFromRoomRequest
    ) -> RusotoFuture<DisassociateSkillGroupFromRoomResponse, DisassociateSkillGroupFromRoomError>;

    fn forget_smart_home_appliances(
        &self, 
        input: ForgetSmartHomeAppliancesRequest
    ) -> RusotoFuture<ForgetSmartHomeAppliancesResponse, ForgetSmartHomeAppliancesError>;

    fn get_address_book(
        &self, 
        input: GetAddressBookRequest
    ) -> RusotoFuture<GetAddressBookResponse, GetAddressBookError>;

    fn get_conference_preference(
        &self
    ) -> RusotoFuture<GetConferencePreferenceResponse, GetConferencePreferenceError>;

    fn get_conference_provider(
        &self, 
        input: GetConferenceProviderRequest
    ) -> RusotoFuture<GetConferenceProviderResponse, GetConferenceProviderError>;

    fn get_contact(
        &self, 
        input: GetContactRequest
    ) -> RusotoFuture<GetContactResponse, GetContactError>;

    fn get_device(
        &self, 
        input: GetDeviceRequest
    ) -> RusotoFuture<GetDeviceResponse, GetDeviceError>;

    fn get_profile(
        &self, 
        input: GetProfileRequest
    ) -> RusotoFuture<GetProfileResponse, GetProfileError>;

    fn get_room(
        &self, 
        input: GetRoomRequest
    ) -> RusotoFuture<GetRoomResponse, GetRoomError>;

    fn get_room_skill_parameter(
        &self, 
        input: GetRoomSkillParameterRequest
    ) -> RusotoFuture<GetRoomSkillParameterResponse, GetRoomSkillParameterError>;

    fn get_skill_group(
        &self, 
        input: GetSkillGroupRequest
    ) -> RusotoFuture<GetSkillGroupResponse, GetSkillGroupError>;

    fn list_business_report_schedules(
        &self, 
        input: ListBusinessReportSchedulesRequest
    ) -> RusotoFuture<ListBusinessReportSchedulesResponse, ListBusinessReportSchedulesError>;

    fn list_conference_providers(
        &self, 
        input: ListConferenceProvidersRequest
    ) -> RusotoFuture<ListConferenceProvidersResponse, ListConferenceProvidersError>;

    fn list_device_events(
        &self, 
        input: ListDeviceEventsRequest
    ) -> RusotoFuture<ListDeviceEventsResponse, ListDeviceEventsError>;

    fn list_skills(
        &self, 
        input: ListSkillsRequest
    ) -> RusotoFuture<ListSkillsResponse, ListSkillsError>;

    fn list_skills_store_categories(
        &self, 
        input: ListSkillsStoreCategoriesRequest
    ) -> RusotoFuture<ListSkillsStoreCategoriesResponse, ListSkillsStoreCategoriesError>;

    fn list_skills_store_skills_by_category(
        &self, 
        input: ListSkillsStoreSkillsByCategoryRequest
    ) -> RusotoFuture<ListSkillsStoreSkillsByCategoryResponse, ListSkillsStoreSkillsByCategoryError>;

    fn list_smart_home_appliances(
        &self, 
        input: ListSmartHomeAppliancesRequest
    ) -> RusotoFuture<ListSmartHomeAppliancesResponse, ListSmartHomeAppliancesError>;

    fn list_tags(
        &self, 
        input: ListTagsRequest
    ) -> RusotoFuture<ListTagsResponse, ListTagsError>;

    fn put_conference_preference(
        &self, 
        input: PutConferencePreferenceRequest
    ) -> RusotoFuture<PutConferencePreferenceResponse, PutConferencePreferenceError>;

    fn put_room_skill_parameter(
        &self, 
        input: PutRoomSkillParameterRequest
    ) -> RusotoFuture<PutRoomSkillParameterResponse, PutRoomSkillParameterError>;

    fn put_skill_authorization(
        &self, 
        input: PutSkillAuthorizationRequest
    ) -> RusotoFuture<PutSkillAuthorizationResponse, PutSkillAuthorizationError>;

    fn register_avs_device(
        &self, 
        input: RegisterAVSDeviceRequest
    ) -> RusotoFuture<RegisterAVSDeviceResponse, RegisterAVSDeviceError>;

    fn reject_skill(
        &self, 
        input: RejectSkillRequest
    ) -> RusotoFuture<RejectSkillResponse, RejectSkillError>;

    fn resolve_room(
        &self, 
        input: ResolveRoomRequest
    ) -> RusotoFuture<ResolveRoomResponse, ResolveRoomError>;

    fn revoke_invitation(
        &self, 
        input: RevokeInvitationRequest
    ) -> RusotoFuture<RevokeInvitationResponse, RevokeInvitationError>;

    fn search_address_books(
        &self, 
        input: SearchAddressBooksRequest
    ) -> RusotoFuture<SearchAddressBooksResponse, SearchAddressBooksError>;

    fn search_contacts(
        &self, 
        input: SearchContactsRequest
    ) -> RusotoFuture<SearchContactsResponse, SearchContactsError>;

    fn search_devices(
        &self, 
        input: SearchDevicesRequest
    ) -> RusotoFuture<SearchDevicesResponse, SearchDevicesError>;

    fn search_profiles(
        &self, 
        input: SearchProfilesRequest
    ) -> RusotoFuture<SearchProfilesResponse, SearchProfilesError>;

    fn search_rooms(
        &self, 
        input: SearchRoomsRequest
    ) -> RusotoFuture<SearchRoomsResponse, SearchRoomsError>;

    fn search_skill_groups(
        &self, 
        input: SearchSkillGroupsRequest
    ) -> RusotoFuture<SearchSkillGroupsResponse, SearchSkillGroupsError>;

    fn search_users(
        &self, 
        input: SearchUsersRequest
    ) -> RusotoFuture<SearchUsersResponse, SearchUsersError>;

    fn send_invitation(
        &self, 
        input: SendInvitationRequest
    ) -> RusotoFuture<SendInvitationResponse, SendInvitationError>;

    fn start_device_sync(
        &self, 
        input: StartDeviceSyncRequest
    ) -> RusotoFuture<StartDeviceSyncResponse, StartDeviceSyncError>;

    fn start_smart_home_appliance_discovery(
        &self, 
        input: StartSmartHomeApplianceDiscoveryRequest
    ) -> RusotoFuture<StartSmartHomeApplianceDiscoveryResponse, StartSmartHomeApplianceDiscoveryError>;

    fn tag_resource(
        &self, 
        input: TagResourceRequest
    ) -> RusotoFuture<TagResourceResponse, TagResourceError>;

    fn untag_resource(
        &self, 
        input: UntagResourceRequest
    ) -> RusotoFuture<UntagResourceResponse, UntagResourceError>;

    fn update_address_book(
        &self, 
        input: UpdateAddressBookRequest
    ) -> RusotoFuture<UpdateAddressBookResponse, UpdateAddressBookError>;

    fn update_business_report_schedule(
        &self, 
        input: UpdateBusinessReportScheduleRequest
    ) -> RusotoFuture<UpdateBusinessReportScheduleResponse, UpdateBusinessReportScheduleError>;

    fn update_conference_provider(
        &self, 
        input: UpdateConferenceProviderRequest
    ) -> RusotoFuture<UpdateConferenceProviderResponse, UpdateConferenceProviderError>;

    fn update_contact(
        &self, 
        input: UpdateContactRequest
    ) -> RusotoFuture<UpdateContactResponse, UpdateContactError>;

    fn update_device(
        &self, 
        input: UpdateDeviceRequest
    ) -> RusotoFuture<UpdateDeviceResponse, UpdateDeviceError>;

    fn update_profile(
        &self, 
        input: UpdateProfileRequest
    ) -> RusotoFuture<UpdateProfileResponse, UpdateProfileError>;

    fn update_room(
        &self, 
        input: UpdateRoomRequest
    ) -> RusotoFuture<UpdateRoomResponse, UpdateRoomError>;

    fn update_skill_group(
        &self, 
        input: UpdateSkillGroupRequest
    ) -> RusotoFuture<UpdateSkillGroupResponse, UpdateSkillGroupError>;
}

Trait representing the capabilities of the Alexa For Business API. Alexa For Business clients implement this trait.

Required methods

fn approve_skill(
&self,
input: ApproveSkillRequest
) -> RusotoFuture<ApproveSkillResponse, ApproveSkillError>

Associates a skill with the organization under the customer's AWS account. If a skill is private, the user implicitly accepts access to this skill during enablement.

fn associate_contact_with_address_book(
&self,
input: AssociateContactWithAddressBookRequest
) -> RusotoFuture<AssociateContactWithAddressBookResponse, AssociateContactWithAddressBookError>

Associates a contact with a given address book.

fn associate_device_with_room(
&self,
input: AssociateDeviceWithRoomRequest
) -> RusotoFuture<AssociateDeviceWithRoomResponse, AssociateDeviceWithRoomError>

Associates a device with a given room. This applies all the settings from the room profile to the device, and all the skills in any skill groups added to that room. This operation requires the device to be online, or else a manual sync is required.

fn associate_skill_group_with_room(
&self,
input: AssociateSkillGroupWithRoomRequest
) -> RusotoFuture<AssociateSkillGroupWithRoomResponse, AssociateSkillGroupWithRoomError>

Associates a skill group with a given room. This enables all skills in the associated skill group on all devices in the room.

fn associate_skill_with_skill_group(
&self,
input: AssociateSkillWithSkillGroupRequest
) -> RusotoFuture<AssociateSkillWithSkillGroupResponse, AssociateSkillWithSkillGroupError>

Associates a skill with a skill group.

fn associate_skill_with_users(
&self,
input: AssociateSkillWithUsersRequest
) -> RusotoFuture<AssociateSkillWithUsersResponse, AssociateSkillWithUsersError>

Makes a private skill available for enrolled users to enable on their devices.

fn create_address_book(
&self,
input: CreateAddressBookRequest
) -> RusotoFuture<CreateAddressBookResponse, CreateAddressBookError>

Creates an address book with the specified details.

fn create_business_report_schedule(
&self,
input: CreateBusinessReportScheduleRequest
) -> RusotoFuture<CreateBusinessReportScheduleResponse, CreateBusinessReportScheduleError>

Creates a recurring schedule for usage reports to deliver to the specified S3 location with a specified daily or weekly interval.

fn create_conference_provider(
&self,
input: CreateConferenceProviderRequest
) -> RusotoFuture<CreateConferenceProviderResponse, CreateConferenceProviderError>

Adds a new conference provider under the user's AWS account.

fn create_contact(
&self,
input: CreateContactRequest
) -> RusotoFuture<CreateContactResponse, CreateContactError>

Creates a contact with the specified details.

fn create_profile(
&self,
input: CreateProfileRequest
) -> RusotoFuture<CreateProfileResponse, CreateProfileError>

Creates a new room profile with the specified details.

fn create_room(
&self,
input: CreateRoomRequest
) -> RusotoFuture<CreateRoomResponse, CreateRoomError>

Creates a room with the specified details.

fn create_skill_group(
&self,
input: CreateSkillGroupRequest
) -> RusotoFuture<CreateSkillGroupResponse, CreateSkillGroupError>

Creates a skill group with a specified name and description.

fn create_user(
&self,
input: CreateUserRequest
) -> RusotoFuture<CreateUserResponse, CreateUserError>

Creates a user.

fn delete_address_book(
&self,
input: DeleteAddressBookRequest
) -> RusotoFuture<DeleteAddressBookResponse, DeleteAddressBookError>

Deletes an address book by the address book ARN.

fn delete_business_report_schedule(
&self,
input: DeleteBusinessReportScheduleRequest
) -> RusotoFuture<DeleteBusinessReportScheduleResponse, DeleteBusinessReportScheduleError>

Deletes the recurring report delivery schedule with the specified schedule ARN.

fn delete_conference_provider(
&self,
input: DeleteConferenceProviderRequest
) -> RusotoFuture<DeleteConferenceProviderResponse, DeleteConferenceProviderError>

Deletes a conference provider.

fn delete_contact(
&self,
input: DeleteContactRequest
) -> RusotoFuture<DeleteContactResponse, DeleteContactError>

Deletes a contact by the contact ARN.

fn delete_device(
&self,
input: DeleteDeviceRequest
) -> RusotoFuture<DeleteDeviceResponse, DeleteDeviceError>

Removes a device from Alexa For Business.

fn delete_profile(
&self,
input: DeleteProfileRequest
) -> RusotoFuture<DeleteProfileResponse, DeleteProfileError>

Deletes a room profile by the profile ARN.

fn delete_room(
&self,
input: DeleteRoomRequest
) -> RusotoFuture<DeleteRoomResponse, DeleteRoomError>

Deletes a room by the room ARN.

fn delete_room_skill_parameter(
&self,
input: DeleteRoomSkillParameterRequest
) -> RusotoFuture<DeleteRoomSkillParameterResponse, DeleteRoomSkillParameterError>

Deletes room skill parameter details by room, skill, and parameter key ID.

fn delete_skill_authorization(
&self,
input: DeleteSkillAuthorizationRequest
) -> RusotoFuture<DeleteSkillAuthorizationResponse, DeleteSkillAuthorizationError>

Unlinks a third-party account from a skill.

fn delete_skill_group(
&self,
input: DeleteSkillGroupRequest
) -> RusotoFuture<DeleteSkillGroupResponse, DeleteSkillGroupError>

Deletes a skill group by skill group ARN.

fn delete_user(
&self,
input: DeleteUserRequest
) -> RusotoFuture<DeleteUserResponse, DeleteUserError>

Deletes a specified user by user ARN and enrollment ARN.

fn disassociate_contact_from_address_book(
&self,
input: DisassociateContactFromAddressBookRequest
) -> RusotoFuture<DisassociateContactFromAddressBookResponse, DisassociateContactFromAddressBookError>

Disassociates a contact from a given address book.

fn disassociate_device_from_room(
&self,
input: DisassociateDeviceFromRoomRequest
) -> RusotoFuture<DisassociateDeviceFromRoomResponse, DisassociateDeviceFromRoomError>

Disassociates a device from its current room. The device continues to be connected to the Wi-Fi network and is still registered to the account. The device settings and skills are removed from the room.

fn disassociate_skill_from_skill_group(
&self,
input: DisassociateSkillFromSkillGroupRequest
) -> RusotoFuture<DisassociateSkillFromSkillGroupResponse, DisassociateSkillFromSkillGroupError>

Disassociates a skill from a skill group.

fn disassociate_skill_from_users(
&self,
input: DisassociateSkillFromUsersRequest
) -> RusotoFuture<DisassociateSkillFromUsersResponse, DisassociateSkillFromUsersError>

Makes a private skill unavailable for enrolled users and prevents them from enabling it on their devices.

fn disassociate_skill_group_from_room(
&self,
input: DisassociateSkillGroupFromRoomRequest
) -> RusotoFuture<DisassociateSkillGroupFromRoomResponse, DisassociateSkillGroupFromRoomError>

Disassociates a skill group from a specified room. This disables all skills in the skill group on all devices in the room.

fn forget_smart_home_appliances(
&self,
input: ForgetSmartHomeAppliancesRequest
) -> RusotoFuture<ForgetSmartHomeAppliancesResponse, ForgetSmartHomeAppliancesError>

Forgets smart home appliances associated to a room.

fn get_address_book(
&self,
input: GetAddressBookRequest
) -> RusotoFuture<GetAddressBookResponse, GetAddressBookError>

Gets address the book details by the address book ARN.

fn get_conference_preference(
&self
) -> RusotoFuture<GetConferencePreferenceResponse, GetConferencePreferenceError>

Retrieves the existing conference preferences.

fn get_conference_provider(
&self,
input: GetConferenceProviderRequest
) -> RusotoFuture<GetConferenceProviderResponse, GetConferenceProviderError>

Gets details about a specific conference provider.

fn get_contact(
&self,
input: GetContactRequest
) -> RusotoFuture<GetContactResponse, GetContactError>

Gets the contact details by the contact ARN.

fn get_device(
&self,
input: GetDeviceRequest
) -> RusotoFuture<GetDeviceResponse, GetDeviceError>

Gets the details of a device by device ARN.

fn get_profile(
&self,
input: GetProfileRequest
) -> RusotoFuture<GetProfileResponse, GetProfileError>

Gets the details of a room profile by profile ARN.

fn get_room(
&self,
input: GetRoomRequest
) -> RusotoFuture<GetRoomResponse, GetRoomError>

Gets room details by room ARN.

fn get_room_skill_parameter(
&self,
input: GetRoomSkillParameterRequest
) -> RusotoFuture<GetRoomSkillParameterResponse, GetRoomSkillParameterError>

Gets room skill parameter details by room, skill, and parameter key ARN.

fn get_skill_group(
&self,
input: GetSkillGroupRequest
) -> RusotoFuture<GetSkillGroupResponse, GetSkillGroupError>

Gets skill group details by skill group ARN.

fn list_business_report_schedules(
&self,
input: ListBusinessReportSchedulesRequest
) -> RusotoFuture<ListBusinessReportSchedulesResponse, ListBusinessReportSchedulesError>

Lists the details of the schedules that a user configured.

fn list_conference_providers(
&self,
input: ListConferenceProvidersRequest
) -> RusotoFuture<ListConferenceProvidersResponse, ListConferenceProvidersError>

Lists conference providers under a specific AWS account.

fn list_device_events(
&self,
input: ListDeviceEventsRequest
) -> RusotoFuture<ListDeviceEventsResponse, ListDeviceEventsError>

Lists the device event history, including device connection status, for up to 30 days.

fn list_skills(
&self,
input: ListSkillsRequest
) -> RusotoFuture<ListSkillsResponse, ListSkillsError>

Lists all enabled skills in a specific skill group.

fn list_skills_store_categories(
&self,
input: ListSkillsStoreCategoriesRequest
) -> RusotoFuture<ListSkillsStoreCategoriesResponse, ListSkillsStoreCategoriesError>

Lists all categories in the Alexa skill store.

fn list_skills_store_skills_by_category(
&self,
input: ListSkillsStoreSkillsByCategoryRequest
) -> RusotoFuture<ListSkillsStoreSkillsByCategoryResponse, ListSkillsStoreSkillsByCategoryError>

Lists all skills in the Alexa skill store by category.

fn list_smart_home_appliances(
&self,
input: ListSmartHomeAppliancesRequest
) -> RusotoFuture<ListSmartHomeAppliancesResponse, ListSmartHomeAppliancesError>

Lists all of the smart home appliances associated with a room.

fn list_tags(
&self,
input: ListTagsRequest
) -> RusotoFuture<ListTagsResponse, ListTagsError>

Lists all tags for the specified resource.

fn put_conference_preference(
&self,
input: PutConferencePreferenceRequest
) -> RusotoFuture<PutConferencePreferenceResponse, PutConferencePreferenceError>

Sets the conference preferences on a specific conference provider at the account level.

fn put_room_skill_parameter(
&self,
input: PutRoomSkillParameterRequest
) -> RusotoFuture<PutRoomSkillParameterResponse, PutRoomSkillParameterError>

Updates room skill parameter details by room, skill, and parameter key ID. Not all skills have a room skill parameter.

fn put_skill_authorization(
&self,
input: PutSkillAuthorizationRequest
) -> RusotoFuture<PutSkillAuthorizationResponse, PutSkillAuthorizationError>

Links a user's account to a third-party skill provider. If this API operation is called by an assumed IAM role, the skill being linked must be a private skill. Also, the skill must be owned by the AWS account that assumed the IAM role.

fn register_avs_device(
&self,
input: RegisterAVSDeviceRequest
) -> RusotoFuture<RegisterAVSDeviceResponse, RegisterAVSDeviceError>

Registers an Alexa-enabled device built by an Original Equipment Manufacturer (OEM) using Alexa Voice Service (AVS).

fn reject_skill(
&self,
input: RejectSkillRequest
) -> RusotoFuture<RejectSkillResponse, RejectSkillError>

Disassociates a skill from the organization under a user's AWS account. If the skill is a private skill, it moves to an AcceptStatus of PENDING. Any private or public skill that is rejected can be added later by calling the ApproveSkill API.

fn resolve_room(
&self,
input: ResolveRoomRequest
) -> RusotoFuture<ResolveRoomResponse, ResolveRoomError>

Determines the details for the room from which a skill request was invoked. This operation is used by skill developers.

fn revoke_invitation(
&self,
input: RevokeInvitationRequest
) -> RusotoFuture<RevokeInvitationResponse, RevokeInvitationError>

Revokes an invitation and invalidates the enrollment URL.

fn search_address_books(
&self,
input: SearchAddressBooksRequest
) -> RusotoFuture<SearchAddressBooksResponse, SearchAddressBooksError>

Searches address books and lists the ones that meet a set of filter and sort criteria.

fn search_contacts(
&self,
input: SearchContactsRequest
) -> RusotoFuture<SearchContactsResponse, SearchContactsError>

Searches contacts and lists the ones that meet a set of filter and sort criteria.

fn search_devices(
&self,
input: SearchDevicesRequest
) -> RusotoFuture<SearchDevicesResponse, SearchDevicesError>

Searches devices and lists the ones that meet a set of filter criteria.

fn search_profiles(
&self,
input: SearchProfilesRequest
) -> RusotoFuture<SearchProfilesResponse, SearchProfilesError>

Searches room profiles and lists the ones that meet a set of filter criteria.

fn search_rooms(
&self,
input: SearchRoomsRequest
) -> RusotoFuture<SearchRoomsResponse, SearchRoomsError>

Searches rooms and lists the ones that meet a set of filter and sort criteria.

fn search_skill_groups(
&self,
input: SearchSkillGroupsRequest
) -> RusotoFuture<SearchSkillGroupsResponse, SearchSkillGroupsError>

Searches skill groups and lists the ones that meet a set of filter and sort criteria.

fn search_users(
&self,
input: SearchUsersRequest
) -> RusotoFuture<SearchUsersResponse, SearchUsersError>

Searches users and lists the ones that meet a set of filter and sort criteria.

fn send_invitation(
&self,
input: SendInvitationRequest
) -> RusotoFuture<SendInvitationResponse, SendInvitationError>

Sends an enrollment invitation email with a URL to a user. The URL is valid for 72 hours or until you call this operation again, whichever comes first.

fn start_device_sync(
&self,
input: StartDeviceSyncRequest
) -> RusotoFuture<StartDeviceSyncResponse, StartDeviceSyncError>

Resets a device and its account to the known default settings, by clearing all information and settings set by previous users.

fn start_smart_home_appliance_discovery(
&self,
input: StartSmartHomeApplianceDiscoveryRequest
) -> RusotoFuture<StartSmartHomeApplianceDiscoveryResponse, StartSmartHomeApplianceDiscoveryError>

Initiates the discovery of any smart home appliances associated with the room.

fn tag_resource(
&self,
input: TagResourceRequest
) -> RusotoFuture<TagResourceResponse, TagResourceError>

Adds metadata tags to a specified resource.

fn untag_resource(
&self,
input: UntagResourceRequest
) -> RusotoFuture<UntagResourceResponse, UntagResourceError>

Removes metadata tags from a specified resource.

fn update_address_book(
&self,
input: UpdateAddressBookRequest
) -> RusotoFuture<UpdateAddressBookResponse, UpdateAddressBookError>

Updates address book details by the address book ARN.

fn update_business_report_schedule(
&self,
input: UpdateBusinessReportScheduleRequest
) -> RusotoFuture<UpdateBusinessReportScheduleResponse, UpdateBusinessReportScheduleError>

Updates the configuration of the report delivery schedule with the specified schedule ARN.

fn update_conference_provider(
&self,
input: UpdateConferenceProviderRequest
) -> RusotoFuture<UpdateConferenceProviderResponse, UpdateConferenceProviderError>

Updates an existing conference provider's settings.

fn update_contact(
&self,
input: UpdateContactRequest
) -> RusotoFuture<UpdateContactResponse, UpdateContactError>

Updates the contact details by the contact ARN.

fn update_device(
&self,
input: UpdateDeviceRequest
) -> RusotoFuture<UpdateDeviceResponse, UpdateDeviceError>

Updates the device name by device ARN.

fn update_profile(
&self,
input: UpdateProfileRequest
) -> RusotoFuture<UpdateProfileResponse, UpdateProfileError>

Updates an existing room profile by room profile ARN.

fn update_room(
&self,
input: UpdateRoomRequest
) -> RusotoFuture<UpdateRoomResponse, UpdateRoomError>

Updates room details by room ARN.

fn update_skill_group(
&self,
input: UpdateSkillGroupRequest
) -> RusotoFuture<UpdateSkillGroupResponse, UpdateSkillGroupError>

Updates skill group details by skill group ARN.

Loading content...