Crate rusoto_alexaforbusiness

Alexa for Business helps you use Alexa in your organization. Alexa for Business provides you with the tools to manage Alexa devices, enroll your users, and assign skills, at scale. You can build your own context-aware voice skills using the Alexa Skills Kit and the Alexa for Business API operations. You can also make these available as private skills for your organization. Alexa for Business makes it efficient to voice-enable your products and services, thus providing context-aware voice experiences for your customers. Device makers building with the Alexa Voice Service (AVS) can create fully integrated solutions, register their products with Alexa for Business, and manage them as shared devices in their organization.

If you're using the service, you're probably looking for AlexaForBusinessClient and AlexaForBusiness.

Structs

AddressBook
An address book with attributes.

AddressBookData
Information related to an address book.

AlexaForBusinessClient
A client for the Alexa For Business API.

ApproveSkillRequest

ApproveSkillResponse

AssociateContactWithAddressBookRequest

AssociateContactWithAddressBookResponse

AssociateDeviceWithNetworkProfileRequest

AssociateDeviceWithNetworkProfileResponse

AssociateDeviceWithRoomRequest

AssociateDeviceWithRoomResponse

AssociateSkillGroupWithRoomRequest

AssociateSkillGroupWithRoomResponse

AssociateSkillWithSkillGroupRequest

AssociateSkillWithSkillGroupResponse

AssociateSkillWithUsersRequest

AssociateSkillWithUsersResponse

Audio
The audio message. There is a 1 MB limit on the audio file input and the only supported format is MP3. To convert your MP3 audio files to an Alexa-friendly,

required codec version (MPEG version 2) and bit rate (48 kbps), you might use converter software. One option for this is a command-line tool, FFmpeg. For more information, see FFmpeg. The following command converts the provided <input-file> to an MP3 file that is played in the announcement:

ffmpeg -i <input-file> -ac 2 -codec:a libmp3lame -b:a 48k -ar 16000 <output-file.mp3>

BusinessReport
Usage report with specified parameters.

BusinessReportContentRange
The content range of the report.

BusinessReportRecurrence
The recurrence of the reports.

BusinessReportS3Location
The S3 location of the output reports.

BusinessReportSchedule
The schedule of the usage report.

Category
The skill store category that is shown. Alexa skills are assigned a specific skill category during creation, such as News, Social, and Sports.

ConferencePreference
The default conference provider that is used if no other scheduled meetings are detected.

ConferenceProvider
An entity that provides a conferencing solution. Alexa for Business acts as the voice interface and mediator that connects users to their preferred conference provider. Examples of conference providers include Amazon Chime, Zoom, Cisco, and Polycom.

Contact
A contact with attributes.

ContactData
Information related to a contact.

Content
The content definition. This can contain only one text, SSML, or audio list object.

CreateAddressBookRequest

CreateAddressBookResponse

CreateBusinessReportScheduleRequest

CreateBusinessReportScheduleResponse

CreateConferenceProviderRequest

CreateConferenceProviderResponse

CreateContactRequest

CreateContactResponse

CreateGatewayGroupRequest

CreateGatewayGroupResponse

CreateNetworkProfileRequest

CreateNetworkProfileResponse

CreateProfileRequest

CreateProfileResponse

CreateRoomRequest

CreateRoomResponse

CreateSkillGroupRequest

CreateSkillGroupResponse

CreateUserRequest

CreateUserResponse

DeleteAddressBookRequest

DeleteAddressBookResponse

DeleteBusinessReportScheduleRequest

DeleteBusinessReportScheduleResponse

DeleteConferenceProviderRequest

DeleteConferenceProviderResponse

DeleteContactRequest

DeleteContactResponse

DeleteDeviceRequest

DeleteDeviceResponse

DeleteDeviceUsageDataRequest

DeleteDeviceUsageDataResponse

DeleteGatewayGroupRequest

DeleteGatewayGroupResponse

DeleteNetworkProfileRequest

DeleteNetworkProfileResponse

DeleteProfileRequest

DeleteProfileResponse

DeleteRoomRequest

DeleteRoomResponse

DeleteRoomSkillParameterRequest

DeleteRoomSkillParameterResponse

DeleteSkillAuthorizationRequest

DeleteSkillAuthorizationResponse

DeleteSkillGroupRequest

DeleteSkillGroupResponse

DeleteUserRequest

DeleteUserResponse

DeveloperInfo
The details about the developer that published the skill.

Device
A device with attributes.

DeviceData
Device attributes.

DeviceEvent
The list of device events.

DeviceNetworkProfileInfo
Detailed information about a device's network profile.

DeviceStatusDetail
Details of a device’s status.

DeviceStatusInfo
Detailed information about a device's status.

DisassociateContactFromAddressBookRequest

DisassociateContactFromAddressBookResponse

DisassociateDeviceFromRoomRequest

DisassociateDeviceFromRoomResponse

DisassociateSkillFromSkillGroupRequest

DisassociateSkillFromSkillGroupResponse

DisassociateSkillFromUsersRequest

DisassociateSkillFromUsersResponse

DisassociateSkillGroupFromRoomRequest

DisassociateSkillGroupFromRoomResponse

Filter
A filter name and value pair that is used to return a more specific list of results. Filters can be used to match a set of resources by various criteria.

ForgetSmartHomeAppliancesRequest

ForgetSmartHomeAppliancesResponse

Gateway
The details of the gateway.

GatewayGroup
The details of the gateway group.

GatewayGroupSummary
The summary of a gateway group.

GatewaySummary
The summary of a gateway.

GetAddressBookRequest

GetAddressBookResponse

GetConferencePreferenceRequest

GetConferencePreferenceResponse

GetConferenceProviderRequest

GetConferenceProviderResponse

GetContactRequest

GetContactResponse

GetDeviceRequest

GetDeviceResponse

GetGatewayGroupRequest

GetGatewayGroupResponse

GetGatewayRequest

GetGatewayResponse

GetInvitationConfigurationRequest

GetInvitationConfigurationResponse

GetNetworkProfileRequest

GetNetworkProfileResponse

GetProfileRequest

GetProfileResponse

GetRoomRequest

GetRoomResponse

GetRoomSkillParameterRequest

GetRoomSkillParameterResponse

GetSkillGroupRequest

GetSkillGroupResponse

IPDialIn
The IP endpoint and protocol for calling.

ListBusinessReportSchedulesRequest

ListBusinessReportSchedulesResponse

ListConferenceProvidersRequest

ListConferenceProvidersResponse

ListDeviceEventsRequest

ListDeviceEventsResponse

ListGatewayGroupsRequest

ListGatewayGroupsResponse

ListGatewaysRequest

ListGatewaysResponse

ListSkillsRequest

ListSkillsResponse

ListSkillsStoreCategoriesRequest

ListSkillsStoreCategoriesResponse

ListSkillsStoreSkillsByCategoryRequest

ListSkillsStoreSkillsByCategoryResponse

ListSmartHomeAppliancesRequest

ListSmartHomeAppliancesResponse

ListTagsRequest

ListTagsResponse

MeetingSetting

The values that indicate whether a pin is always required (YES), never required (NO), or OPTIONAL.

If YES, Alexa will always ask for a meeting pin.

If NO, Alexa will never ask for a meeting pin.

If OPTIONAL, Alexa will ask if you have a meeting pin and if the customer responds with yes, it will ask for the meeting pin.

NetworkProfile
The network profile associated with a device.

NetworkProfileData
The data associated with a network profile.

PSTNDialIn
The information for public switched telephone network (PSTN) conferencing.

Profile
A room profile with attributes.

ProfileData
The data of a room profile.

PutConferencePreferenceRequest

PutConferencePreferenceResponse

PutInvitationConfigurationRequest

PutInvitationConfigurationResponse

PutRoomSkillParameterRequest

PutRoomSkillParameterResponse

PutSkillAuthorizationRequest

PutSkillAuthorizationResponse

RegisterAVSDeviceRequest

RegisterAVSDeviceResponse

RejectSkillRequest

RejectSkillResponse

ResolveRoomRequest

ResolveRoomResponse

RevokeInvitationRequest

RevokeInvitationResponse

Room
A room with attributes.

RoomData
The data of a room.

RoomSkillParameter
A skill parameter associated with a room.

SearchAddressBooksRequest

SearchAddressBooksResponse

SearchContactsRequest

SearchContactsResponse

SearchDevicesRequest

SearchDevicesResponse

SearchNetworkProfilesRequest

SearchNetworkProfilesResponse

SearchProfilesRequest

SearchProfilesResponse

SearchRoomsRequest

SearchRoomsResponse

SearchSkillGroupsRequest

SearchSkillGroupsResponse

SearchUsersRequest

SearchUsersResponse

SendAnnouncementRequest

SendAnnouncementResponse

SendInvitationRequest

SendInvitationResponse

SkillDetails
Granular information about the skill.

SkillGroup
A skill group with attributes.

SkillGroupData
The attributes of a skill group.

SkillSummary
The summary of skills.

SkillsStoreSkill
The detailed information about an Alexa skill.

SmartHomeAppliance
A smart home appliance that can connect to a central system. Any domestic device can be a smart appliance.

Sort
An object representing a sort criteria.

Ssml
The SSML message. For more information, see SSML Reference.

StartDeviceSyncRequest

StartDeviceSyncResponse

StartSmartHomeApplianceDiscoveryRequest

StartSmartHomeApplianceDiscoveryResponse

Tag
A key-value pair that can be associated with a resource.

TagResourceRequest

TagResourceResponse

Text
The text message.

UntagResourceRequest

UntagResourceResponse

UpdateAddressBookRequest

UpdateAddressBookResponse

UpdateBusinessReportScheduleRequest

UpdateBusinessReportScheduleResponse

UpdateConferenceProviderRequest

UpdateConferenceProviderResponse

UpdateContactRequest

UpdateContactResponse

UpdateDeviceRequest

UpdateDeviceResponse

UpdateGatewayGroupRequest

UpdateGatewayGroupResponse

UpdateGatewayRequest

UpdateGatewayResponse

UpdateNetworkProfileRequest

UpdateNetworkProfileResponse

UpdateProfileRequest

UpdateProfileResponse

UpdateRoomRequest

UpdateRoomResponse

UpdateSkillGroupRequest

UpdateSkillGroupResponse

UserData
Information related to a user.

Enums

ApproveSkillError
Errors returned by ApproveSkill

AssociateContactWithAddressBookError
Errors returned by AssociateContactWithAddressBook

AssociateDeviceWithNetworkProfileError
Errors returned by AssociateDeviceWithNetworkProfile

AssociateDeviceWithRoomError
Errors returned by AssociateDeviceWithRoom

AssociateSkillGroupWithRoomError
Errors returned by AssociateSkillGroupWithRoom

AssociateSkillWithSkillGroupError
Errors returned by AssociateSkillWithSkillGroup

AssociateSkillWithUsersError
Errors returned by AssociateSkillWithUsers

CreateAddressBookError
Errors returned by CreateAddressBook

CreateBusinessReportScheduleError
Errors returned by CreateBusinessReportSchedule

CreateConferenceProviderError
Errors returned by CreateConferenceProvider

CreateContactError
Errors returned by CreateContact

CreateGatewayGroupError
Errors returned by CreateGatewayGroup

CreateNetworkProfileError
Errors returned by CreateNetworkProfile

CreateProfileError
Errors returned by CreateProfile

CreateRoomError
Errors returned by CreateRoom

CreateSkillGroupError
Errors returned by CreateSkillGroup

CreateUserError
Errors returned by CreateUser

DeleteAddressBookError
Errors returned by DeleteAddressBook

DeleteBusinessReportScheduleError
Errors returned by DeleteBusinessReportSchedule

DeleteConferenceProviderError
Errors returned by DeleteConferenceProvider

DeleteContactError
Errors returned by DeleteContact

DeleteDeviceError
Errors returned by DeleteDevice

DeleteDeviceUsageDataError
Errors returned by DeleteDeviceUsageData

DeleteGatewayGroupError
Errors returned by DeleteGatewayGroup

DeleteNetworkProfileError
Errors returned by DeleteNetworkProfile

DeleteProfileError
Errors returned by DeleteProfile

DeleteRoomError
Errors returned by DeleteRoom

DeleteRoomSkillParameterError
Errors returned by DeleteRoomSkillParameter

DeleteSkillAuthorizationError
Errors returned by DeleteSkillAuthorization

DeleteSkillGroupError
Errors returned by DeleteSkillGroup

DeleteUserError
Errors returned by DeleteUser

DisassociateContactFromAddressBookError
Errors returned by DisassociateContactFromAddressBook

DisassociateDeviceFromRoomError
Errors returned by DisassociateDeviceFromRoom

DisassociateSkillFromSkillGroupError
Errors returned by DisassociateSkillFromSkillGroup

DisassociateSkillFromUsersError
Errors returned by DisassociateSkillFromUsers

DisassociateSkillGroupFromRoomError
Errors returned by DisassociateSkillGroupFromRoom

ForgetSmartHomeAppliancesError
Errors returned by ForgetSmartHomeAppliances

GetAddressBookError
Errors returned by GetAddressBook

GetConferencePreferenceError
Errors returned by GetConferencePreference

GetConferenceProviderError
Errors returned by GetConferenceProvider

GetContactError
Errors returned by GetContact

GetDeviceError
Errors returned by GetDevice

GetGatewayError
Errors returned by GetGateway

GetGatewayGroupError
Errors returned by GetGatewayGroup

GetInvitationConfigurationError
Errors returned by GetInvitationConfiguration

GetNetworkProfileError
Errors returned by GetNetworkProfile

GetProfileError
Errors returned by GetProfile

GetRoomError
Errors returned by GetRoom

GetRoomSkillParameterError
Errors returned by GetRoomSkillParameter

GetSkillGroupError
Errors returned by GetSkillGroup

ListBusinessReportSchedulesError
Errors returned by ListBusinessReportSchedules

ListConferenceProvidersError
Errors returned by ListConferenceProviders

ListDeviceEventsError
Errors returned by ListDeviceEvents

ListGatewayGroupsError
Errors returned by ListGatewayGroups

ListGatewaysError
Errors returned by ListGateways

ListSkillsError
Errors returned by ListSkills

ListSkillsStoreCategoriesError
Errors returned by ListSkillsStoreCategories

ListSkillsStoreSkillsByCategoryError
Errors returned by ListSkillsStoreSkillsByCategory

ListSmartHomeAppliancesError
Errors returned by ListSmartHomeAppliances

ListTagsError
Errors returned by ListTags

PutConferencePreferenceError
Errors returned by PutConferencePreference

PutInvitationConfigurationError
Errors returned by PutInvitationConfiguration

PutRoomSkillParameterError
Errors returned by PutRoomSkillParameter

PutSkillAuthorizationError
Errors returned by PutSkillAuthorization

RegisterAVSDeviceError
Errors returned by RegisterAVSDevice

RejectSkillError
Errors returned by RejectSkill

ResolveRoomError
Errors returned by ResolveRoom

RevokeInvitationError
Errors returned by RevokeInvitation

SearchAddressBooksError
Errors returned by SearchAddressBooks

SearchContactsError
Errors returned by SearchContacts

SearchDevicesError
Errors returned by SearchDevices

SearchNetworkProfilesError
Errors returned by SearchNetworkProfiles

SearchProfilesError
Errors returned by SearchProfiles

SearchRoomsError
Errors returned by SearchRooms

SearchSkillGroupsError
Errors returned by SearchSkillGroups

SearchUsersError
Errors returned by SearchUsers

SendAnnouncementError
Errors returned by SendAnnouncement

SendInvitationError
Errors returned by SendInvitation

StartDeviceSyncError
Errors returned by StartDeviceSync

StartSmartHomeApplianceDiscoveryError
Errors returned by StartSmartHomeApplianceDiscovery

TagResourceError
Errors returned by TagResource

UntagResourceError
Errors returned by UntagResource

UpdateAddressBookError
Errors returned by UpdateAddressBook

UpdateBusinessReportScheduleError
Errors returned by UpdateBusinessReportSchedule

UpdateConferenceProviderError
Errors returned by UpdateConferenceProvider

UpdateContactError
Errors returned by UpdateContact

UpdateDeviceError
Errors returned by UpdateDevice

UpdateGatewayError
Errors returned by UpdateGateway

UpdateGatewayGroupError
Errors returned by UpdateGatewayGroup

UpdateNetworkProfileError
Errors returned by UpdateNetworkProfile

UpdateProfileError
Errors returned by UpdateProfile

UpdateRoomError
Errors returned by UpdateRoom

UpdateSkillGroupError
Errors returned by UpdateSkillGroup

Traits

AlexaForBusiness
Trait representing the capabilities of the Alexa For Business API. Alexa For Business clients implement this trait.