Crate rusoto_alexaforbusiness

Alexa for Business makes it easy for you to use Alexa in your organization. Alexa for Business gives you the tools you need for managing Alexa devices, enroll your users, and assign skills, at scale. You can build your own context-aware voice skills using the Alexa Skills Kit and the Alexa for Business API operations. You can make also these available as private skills for your organization. Alexa for Business makes it easy to voice-enable your products and services, providing context-aware voice experiences for your customers.

If you're using the service, you're probably looking for AlexaForBusinessClient and AlexaForBusiness.

Structs

AddressBook
An address book with attributes.

AddressBookData
Information related to an address book.

AlexaForBusinessClient
A client for the Alexa For Business API.

AssociateContactWithAddressBookRequest

AssociateContactWithAddressBookResponse

AssociateDeviceWithRoomRequest

AssociateDeviceWithRoomResponse

AssociateSkillGroupWithRoomRequest

AssociateSkillGroupWithRoomResponse

Contact
A contact with attributes.

ContactData
Information related to a contact.

CreateAddressBookRequest

CreateAddressBookResponse

CreateContactRequest

CreateContactResponse

CreateProfileRequest

CreateProfileResponse

CreateRoomRequest

CreateRoomResponse

CreateSkillGroupRequest

CreateSkillGroupResponse

CreateUserRequest

CreateUserResponse

DeleteAddressBookRequest

DeleteAddressBookResponse

DeleteContactRequest

DeleteContactResponse

DeleteProfileRequest

DeleteProfileResponse

DeleteRoomRequest

DeleteRoomResponse

DeleteRoomSkillParameterRequest

DeleteRoomSkillParameterResponse

DeleteSkillGroupRequest

DeleteSkillGroupResponse

DeleteUserRequest

DeleteUserResponse

Device
A device with attributes.

DeviceData
Device attributes.

DeviceEvent
The list of device events.

DeviceStatusDetail
Details of a device’s status.

DeviceStatusInfo
Detailed information about a device's status.

DisassociateContactFromAddressBookRequest

DisassociateContactFromAddressBookResponse

DisassociateDeviceFromRoomRequest

DisassociateDeviceFromRoomResponse

DisassociateSkillGroupFromRoomRequest

DisassociateSkillGroupFromRoomResponse

Filter
A filter name and value pair that is used to return a more specific list of results. Filters can be used to match a set of resources by various criteria.

GetAddressBookRequest

GetAddressBookResponse

GetContactRequest

GetContactResponse

GetDeviceRequest

GetDeviceResponse

GetProfileRequest

GetProfileResponse

GetRoomRequest

GetRoomResponse

GetRoomSkillParameterRequest

GetRoomSkillParameterResponse

GetSkillGroupRequest

GetSkillGroupResponse

ListDeviceEventsRequest

ListDeviceEventsResponse

ListSkillsRequest

ListSkillsResponse

ListTagsRequest

ListTagsResponse

Profile
A room profile with attributes.

ProfileData
The data of a room profile.

PutRoomSkillParameterRequest

PutRoomSkillParameterResponse

ResolveRoomRequest

ResolveRoomResponse

RevokeInvitationRequest

RevokeInvitationResponse

Room
A room with attributes.

RoomData
The data of a room.

RoomSkillParameter
A skill parameter associated with a room.

SearchAddressBooksRequest

SearchAddressBooksResponse

SearchContactsRequest

SearchContactsResponse

SearchDevicesRequest

SearchDevicesResponse

SearchProfilesRequest

SearchProfilesResponse

SearchRoomsRequest

SearchRoomsResponse

SearchSkillGroupsRequest

SearchSkillGroupsResponse

SearchUsersRequest

SearchUsersResponse

SendInvitationRequest

SendInvitationResponse

SkillGroup
A skill group with attributes.

SkillGroupData
The attributes of a skill group.

SkillSummary
The summary of skills.

Sort
An object representing a sort criteria.

StartDeviceSyncRequest

StartDeviceSyncResponse

Tag
A key-value pair that can be associated with a resource.

TagResourceRequest

TagResourceResponse

UntagResourceRequest

UntagResourceResponse

UpdateAddressBookRequest

UpdateAddressBookResponse

UpdateContactRequest

UpdateContactResponse

UpdateDeviceRequest

UpdateDeviceResponse

UpdateProfileRequest

UpdateProfileResponse

UpdateRoomRequest

UpdateRoomResponse

UpdateSkillGroupRequest

UpdateSkillGroupResponse

UserData
Information related to a user.

Enums

AssociateContactWithAddressBookError
Errors returned by AssociateContactWithAddressBook

AssociateDeviceWithRoomError
Errors returned by AssociateDeviceWithRoom

AssociateSkillGroupWithRoomError
Errors returned by AssociateSkillGroupWithRoom

CreateAddressBookError
Errors returned by CreateAddressBook

CreateContactError
Errors returned by CreateContact

CreateProfileError
Errors returned by CreateProfile

CreateRoomError
Errors returned by CreateRoom

CreateSkillGroupError
Errors returned by CreateSkillGroup

CreateUserError
Errors returned by CreateUser

DeleteAddressBookError
Errors returned by DeleteAddressBook

DeleteContactError
Errors returned by DeleteContact

DeleteProfileError
Errors returned by DeleteProfile

DeleteRoomError
Errors returned by DeleteRoom

DeleteRoomSkillParameterError
Errors returned by DeleteRoomSkillParameter

DeleteSkillGroupError
Errors returned by DeleteSkillGroup

DeleteUserError
Errors returned by DeleteUser

DisassociateContactFromAddressBookError
Errors returned by DisassociateContactFromAddressBook

DisassociateDeviceFromRoomError
Errors returned by DisassociateDeviceFromRoom

DisassociateSkillGroupFromRoomError
Errors returned by DisassociateSkillGroupFromRoom

GetAddressBookError
Errors returned by GetAddressBook

GetContactError
Errors returned by GetContact

GetDeviceError
Errors returned by GetDevice

GetProfileError
Errors returned by GetProfile

GetRoomError
Errors returned by GetRoom

GetRoomSkillParameterError
Errors returned by GetRoomSkillParameter

GetSkillGroupError
Errors returned by GetSkillGroup

ListDeviceEventsError
Errors returned by ListDeviceEvents

ListSkillsError
Errors returned by ListSkills

ListTagsError
Errors returned by ListTags

PutRoomSkillParameterError
Errors returned by PutRoomSkillParameter

ResolveRoomError
Errors returned by ResolveRoom

RevokeInvitationError
Errors returned by RevokeInvitation

SearchAddressBooksError
Errors returned by SearchAddressBooks

SearchContactsError
Errors returned by SearchContacts

SearchDevicesError
Errors returned by SearchDevices

SearchProfilesError
Errors returned by SearchProfiles

SearchRoomsError
Errors returned by SearchRooms

SearchSkillGroupsError
Errors returned by SearchSkillGroups

SearchUsersError
Errors returned by SearchUsers

SendInvitationError
Errors returned by SendInvitation

StartDeviceSyncError
Errors returned by StartDeviceSync

TagResourceError
Errors returned by TagResource

UntagResourceError
Errors returned by UntagResource

UpdateAddressBookError
Errors returned by UpdateAddressBook

UpdateContactError
Errors returned by UpdateContact

UpdateDeviceError
Errors returned by UpdateDevice

UpdateProfileError
Errors returned by UpdateProfile

UpdateRoomError
Errors returned by UpdateRoom

UpdateSkillGroupError
Errors returned by UpdateSkillGroup

Traits

AlexaForBusiness
Trait representing the capabilities of the Alexa For Business API. Alexa For Business clients implement this trait.