Struct rusoto_alexaforbusiness::AlexaForBusinessClient

pub struct AlexaForBusinessClient { /* fields omitted */ }

A client for the Alexa For Business API.

Methods

impl AlexaForBusinessClient

pub fn new(region: Region) -> AlexaForBusinessClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> AlexaForBusinessClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,